Добавлена библиотека, файлы перемещены в src
