--- RabbitMQClient.cpp
@@ -1,6 +1,0 @@
-#include "RabbitMQClient.h"
-
-RabbitMQClient::RabbitMQClient()
-{
-
-}
