--- RabbitMQClient.h
@@ -1,12 +1,0 @@
-#ifndef RABBITMQCLIENT_H
-#define RABBITMQCLIENT_H
-
-#include "amqpcpp.h"
-
-class RabbitMQClient : public AMQP::ConnectionHandler
-{
-public:
-    RabbitMQClient();
-};
-
-#endif // RABBITMQCLIENT_H
