+++ src/RabbitMQClient.cpp
@@ -1,0 +1,26 @@
+#include "RabbitMQClient.h"
+
+RabbitMQClient::RabbitMQClient()
+{
+
+}
+
+void RabbitMQClient::onData(AMQP::Connection *connection, const char *data, size_t size)
+{
+
+}
+
+void RabbitMQClient::onReady(AMQP::Connection *connection)
+{
+
+}
+
+void RabbitMQClient::onError(AMQP::Connection *connection, const char *message)
+{
+
+}
+
+void RabbitMQClient::onClosed(AMQP::Connection *connection)
+{
+
+}
